+++ shaders/lib/commonFunctions.glsl
@@ -6,11 +6,11 @@
     vec3 GetSunVector() {
         const vec2 sunRotationData = vec2(cos(sunPathRotation * 0.01745329251994), -sin(sunPathRotation * 0.01745329251994));
         #ifdef OVERWORLD
-            float ang = fract(timeAngle - 0.25);
+            mediump float ang = fract(timeAngle - 0.25);
             ang = (ang + (cos(ang * 3.14159265358979) * -0.5 + 0.5 - ang) / 3.0) * 6.28318530717959;
             return normalize2((gbufferModelView * vec4(vec3(-sin(ang), cos(ang) * sunRotationData) * 2000.0, 1.0)).xyz);
         #elif defined END
-            float ang = 0.0;
+            mediump float ang = 0.0;
             return normalize2((gbufferModelView * vec4(vec3(0.0, sunRotationData * 2000.0), 1.0)).xyz);
         #else
             return vec3(0.0);
@@ -18,7 +18,7 @@
     }
 #endif
 
-float GetLuminance(vec3 color) {
+mediump float GetLuminance(vec3 color) {
     return dot(color, vec3(0.299, 0.587, 0.114));
 }
 
@@ -26,18 +26,18 @@ vec3 DoLuminanceCorrection(vec3 color) {
     return color / GetLuminance(color);
 }
 
-float GetBiasFactor(float NdotLM) {
-    float NdotLM2 = NdotLM * NdotLM;
+mediump float GetBiasFactor(float NdotLM) {
+    mediump float NdotLM2 = NdotLM * NdotLM;
     return 1.25 * (1.0 - NdotLM2 * NdotLM2) / NdotLM;
 }
 
-float GetHorizonFactor(float XdotU) {
+mediump float GetHorizonFactor(float XdotU) {
     #ifdef SUN_MOON_HORIZON
-        float horizon = clamp((XdotU + 0.1) * 10.0, 0.0, 1.0);
+        mediump float horizon = clamp((XdotU + 0.1) * 10.0, 0.0, 1.0);
         horizon *= horizon;
         return horizon * horizon * (3.0 - 2.0 * horizon);
     #else
-        float horizon = min(XdotU + 1.0, 1.0);
+        mediump float horizon = min(XdotU + 1.0, 1.0);
         horizon *= horizon;
         horizon *= horizon;
         return horizon * horizon;
@@ -56,7 +56,7 @@ bool CheckForStick(vec3 albedo) {
             CheckForColor(albedo, vec3(137, 103, 39));
 }
 
-float GetMaxColorDif(vec3 color) {
+mediump float GetMaxColorDif(vec3 color) {
     vec3 dif = abs(vec3(color.r - color.g, color.g - color.b, color.r - color.b));
     return max(dif.r, max(dif.g, dif.b));
 }
@@ -64,19 +64,19 @@ float GetMaxColorDif(vec3 color) {
 int min1(int x) {
     return min(x, 1);
 }
-float min1(float x) {
+mediump float min1(float x) {
     return min(x, 1.0);
 }
 int max0(int x) {
     return max(x, 0);
 }
-float max0(float x) {
+mediump float max0(float x) {
     return max(x, 0.0);
 }
 int clamp01(int x) {
     return clamp(x, 0, 1);
 }
-float clamp01(float x) {
+mediump float clamp01(float x) {
     return clamp(x, 0.0, 1.0);
 }
 vec2 clamp01(vec2 x) {
@@ -89,7 +89,7 @@ vec3 clamp01(vec3 x) {
 int pow2(int x) {
     return x * x;
 }
-float pow2(float x) {
+mediump float pow2(float x) {
     return x * x;
 }
 vec2 pow2(vec2 x) {
@@ -105,7 +105,7 @@ vec4 pow2(vec4 x) {
 int pow3(int x) {
     return pow2(x) * x;
 }
-float pow3(float x) {
+mediump float pow3(float x) {
     return pow2(x) * x;
 }
 vec2 pow3(vec2 x) {
@@ -118,7 +118,7 @@ vec4 pow3(vec4 x) {
     return pow2(x) * x;
 }
 
-float pow1_5(float x) { // Faster pow(x, 1.5) approximation (that isn't accurate at all) if x is between 0 and 1
+mediump float pow1_5(float x) { // Faster pow(x, 1.5) approximation (that isn't accurate at all) if x is between 0 and 1
     return x - x * pow2(1.0 - x); // Thanks to SixthSurge
 }
 vec2 pow1_5(vec2 x) {
@@ -131,7 +131,7 @@ vec4 pow1_5(vec4 x) {
     return x - x * pow2(1.0 - x);
 }
 
-float sqrt1(float x) { // Faster sqrt() approximation (that isn't accurate at all) if x is between 0 and 1
+mediump float sqrt1(float x) { // Faster sqrt() approximation (that isn't accurate at all) if x is between 0 and 1
     return x * (2.0 - x); // Thanks to Builderb0y
 }
 vec2 sqrt1(vec2 x) {
@@ -143,7 +143,7 @@ vec3 sqrt1(vec3 x) {
 vec4 sqrt1(vec4 x) {
     return x * (2.0 - x);
 }
-float sqrt2(float x) {
+mediump float sqrt2(float x) {
     x = 1.0 - x;
     x *= x;
     x *= x;
@@ -167,7 +167,7 @@ vec4 sqrt2(vec4 x) {
     x *= x;
     return 1.0 - x;
 }
-float sqrt3(float x) {
+mediump float sqrt3(float x) {
     x = 1.0 - x;
     x *= x;
     x *= x;
@@ -195,7 +195,7 @@ vec4 sqrt3(vec4 x) {
     x *= x;
     return 1.0 - x;
 }
-float sqrt4(float x) {
+mediump float sqrt4(float x) {
     x = 1.0 - x;
     x *= x;
     x *= x;
@@ -228,7 +228,7 @@ vec4 sqrt4(vec4 x) {
     return 1.0 - x;
 }
 
-float smoothstep1(float x) {
+mediump float smoothstep1(float x) {
     return x * x * (3.0 - 2.0 * x);
 }
 vec2 smoothstep1(vec2 x) {
@@ -247,8 +247,8 @@ vec3 rgb2hsv(vec3 c)
     vec4 p = mix(vec4(c.bg, K.wz), vec4(c.gb, K.xy), step(c.b, c.g));
     vec4 q = mix(vec4(p.xyw, c.r), vec4(c.r, p.yzx), step(p.x, c.r));
 
-    float d = q.x - min(q.w, q.y);
-    float e = 1.0e-10;
+    mediump float d = q.x - min(q.w, q.y);
+    mediump float e = 1.0e-10;
     return vec3(abs(q.z + (q.w - q.y) / (6.0 * d + e)), d / (q.x + e), q.x);
 }
 vec3 hsv2rgb(vec3 c)
