+++ shaders/lib/optimizationFunctions.glsl
@@ -8,7 +8,7 @@ vec4 normalize2(vec4 a) {
     return a * inversesqrt(dot(a,a));
 }
 
-float pow2(float a, float b) {
+mediump float pow2(float a, float b) {
     return exp2(log2(a) * b);
 }
 
@@ -23,7 +23,7 @@ vec4 pow2(vec4 a, vec4 b) {
     return exp2(log2(a) * b);
 }
 
-float mix2(float a, float b, float c) {
+mediump float mix2(float a, float b, float c) {
     return (b - a) * c + a;
 }
